Generate global state transition functions
The functions used to change the global state are
generated automatically at runtime in favor of maintainability.

The state client will keep trying to change the state if
for some reason it fails, within a reasonable timeout, without
restarting the state or calling another function. This way we
don't increase the call stack and it's easier to trace errors.